+++ app/src/commonMain/kotlin/com/maksimowiczm/foodyou/feature/product/DownloadProductScreen.kt
@@ -1,16 +1,22 @@
 package com.maksimowiczm.foodyou.feature.product
 
 import androidx.compose.foundation.text.input.rememberTextFieldState
+import androidx.compose.foundation.text.input.setTextAndPlaceCursorAtEnd
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalUriHandler
 import androidx.lifecycle.Lifecycle
 import androidx.lifecycle.compose.LocalLifecycleOwner
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.lifecycle.repeatOnLifecycle
+import com.maksimowiczm.foodyou.core.ui.utils.LocalClipboardManager
 import com.maksimowiczm.foodyou.feature.product.domain.RemoteProduct
 import com.maksimowiczm.foodyou.feature.product.ui.download.DownloadProductScreen
 import com.maksimowiczm.foodyou.feature.product.ui.download.DownloadProductScreenViewModel
+import foodyou.app.generated.resources.*
+import foodyou.app.generated.resources.Res
+import org.jetbrains.compose.resources.stringResource
 import org.koin.compose.viewmodel.koinViewModel
 
 @Composable
@@ -45,12 +51,23 @@ private fun DownloadProductScreen(
         }
     }
 
+    val clipboardManager = LocalClipboardManager.current
+    val uriHandler = LocalUriHandler.current
+    val openFoodFactsUrl = stringResource(Res.string.link_open_food_facts)
+
     DownloadProductScreen(
         isMutating = isMutating,
         error = error,
         textFieldState = textFieldState,
         onBack = onBack,
         onDownload = { viewModel.onDownload(textFieldState.text.toString()) },
+        onPaste = {
+            clipboardManager
+                .paste()
+                ?.takeIf { it.isNotEmpty() }
+                ?.let { textFieldState.setTextAndPlaceCursorAtEnd(it) }
+        },
+        onOpenFoodFacts = { uriHandler.openUri(openFoodFactsUrl) },
         modifier = modifier
     )
 }
