+++ app/src/commonMain/kotlin/com/maksimowiczm/foodyou/feature/product/ui/download/ActionChips.kt
@@ -12,42 +12,15 @@ import androidx.compose.material3.Icon
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalUriHandler
 import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.unit.dp
-import com.maksimowiczm.foodyou.core.ui.utils.LocalClipboardManager
 import foodyou.app.generated.resources.Res
 import foodyou.app.generated.resources.action_browse_open_food_facts
 import foodyou.app.generated.resources.action_paste_url
-import foodyou.app.generated.resources.link_open_food_facts
 import org.jetbrains.compose.resources.stringResource
 
 @Composable
 internal fun ActionChips(
-    isMutating: Boolean,
-    onPaste: (String) -> Unit,
-    modifier: Modifier = Modifier
-) {
-    val uriHandler = LocalUriHandler.current
-    val openFoodFactsUrl = stringResource(Res.string.link_open_food_facts)
-
-    val clipboardManager = LocalClipboardManager.current
-
-    ActionChips(
-        isMutating = isMutating,
-        onPaste = {
-            clipboardManager
-                .paste()
-                ?.takeIf { it.isNotEmpty() }
-                ?.let { onPaste(it) }
-        },
-        onOpenFoodFacts = { uriHandler.openUri(openFoodFactsUrl) },
-        modifier = modifier
-    )
-}
-
-@Composable
-private fun ActionChips(
     isMutating: Boolean,
     onPaste: () -> Unit,
     onOpenFoodFacts: () -> Unit,
